+++ src/telnyx.ts
@@ -5,10 +5,159 @@ import {exec} from 'child_process';
 import {AppInfo, TelnyxObject} from './Types';
 import * as utils from './utils';
 
-// TODO: convert other resources to ts
+//#region Resources
+import {AccessIpAddress} from './resources/AccessIpAddress';
+import {AccessIpRanges} from './resources/AccessIpRanges';
+import {AccessTokens} from './resources/AccessTokens';
+import {Actions} from './resources/Actions';
+import {ActionsSimCards} from './resources/ActionsSimCards';
+import {ActivateDeactivateBulkCredentials} from './resources/ActivateDeactivateBulkCredentials';
+import {Addresses} from './resources/Addresses';
+import {AdvancedOptinoptout} from './resources/AdvancedOptinoptout';
+import {AI} from './resources/AI';
+import {AuthenticationProviders} from './resources/AuthenticationProviders';
+import {AvailablePhoneNumbers} from './resources/AvailablePhoneNumbers';
 import {Balance} from './resources/Balance';
+import {Billing} from './resources/Billing';
+import {BillingGroups} from './resources/BillingGroups';
+import {Brands} from './resources/Brands';
+import {Bucket} from './resources/Bucket';
+import {BucketUsage} from './resources/BucketUsage';
+import {BulkCreation} from './resources/BulkCreation';
+import {BulkCredentials} from './resources/BulkCredentials';
+import {BulkPhoneNumberCampaigns} from './resources/BulkPhoneNumberCampaigns';
+import {BulkPhoneNumberOperations} from './resources/BulkPhoneNumberOperations';
+import {BulkSoleProprietorCreation} from './resources/BulkSoleProprietorCreation';
+import {BulkTelephonyCredentials} from './resources/BulkTelephonyCredentials';
+import {BusinessIdentity} from './resources/BusinessIdentity';
+import {CallControlApplications} from './resources/CallControlApplications';
+import {CallEvents} from './resources/CallEvents';
+import {CallInformation} from './resources/CallInformation';
+import {CallRecordings} from './resources/CallRecordings';
+import {Calls} from './resources/Calls';
+import {Campaign} from './resources/Campaign';
+import {CampaignBuilder} from './resources/CampaignBuilder';
+import {CdrUsageReports} from './resources/CdrUsageReports';
+import {Channelzones} from './resources/Channelzones';
+import {ClientStateUpdate} from './resources/ClientStateUpdate';
+import {Conferences} from './resources/Conferences';
+import {Connections} from './resources/Connections';
+import {CredentialConnections} from './resources/CredentialConnections';
+import {Credentials} from './resources/Credentials';
+import {CsvDownloads} from './resources/CsvDownloads';
+import {CustomerServiceRecord} from './resources/CustomerServiceRecord';
+import {Debugging} from './resources/Debugging';
+import {DetailRecords} from './resources/DetailRecords';
+import {DialogflowIntegration} from './resources/DialogflowIntegration';
+import {DocumentLinks} from './resources/DocumentLinks';
+import {Documents} from './resources/Documents';
+import {DynamicEmergency} from './resources/DynamicEmergency';
+import {DynamicEmergencyAddresses} from './resources/DynamicEmergencyAddresses';
+import {DynamicEmergencyEndpoints} from './resources/DynamicEmergencyEndpoints';
+import {Events} from './resources/Events';
+import {ExternalConnections} from './resources/ExternalConnections';
+import {FaxApplications} from './resources/FaxApplications';
+import {Faxes} from './resources/Faxes';
+import {FqdnConnections} from './resources/FqdnConnections';
+import {Fqdns} from './resources/Fqdns';
+import {GlobalIps} from './resources/GlobalIps';
+import {InboundChannels} from './resources/InboundChannels';
+import {InventoryCoverage} from './resources/InventoryCoverage';
+import {InventoryLevel} from './resources/InventoryLevel';
+import {IpAddresses} from './resources/IpAddresses';
+import {IpConnections} from './resources/IpConnections';
+import {IpRanges} from './resources/IpRanges';
+import {Ips} from './resources/Ips';
+import {ManagedAccounts} from './resources/ManagedAccounts';
+import {MdrDetailReports} from './resources/MdrDetailReports';
+import {MdrUsageReports} from './resources/MdrUsageReports';
+import {MediaStorageApi} from './resources/MediaStorageApi';
+import {Messages} from './resources/Messages';
+import {MessagesAlphanumericSenderId} from './resources/MessagesAlphanumericSenderId';
+import {MessagingHostedNumber} from './resources/MessagingHostedNumber';
+import {MessagingHostedNumberOrders} from './resources/MessagingHostedNumberOrders';
+import {MessagingHostedNumbers} from './resources/MessagingHostedNumbers';
+import {MessagingPhoneNumbers} from './resources/MessagingPhoneNumbers';
 import {MessagingProfileMetrics} from './resources/MessagingProfileMetrics';
 import {MessagingProfiles} from './resources/MessagingProfiles';
+import {MessagingSenderIds} from './resources/MessagingSenderIds';
+import {MessagingShortCodes} from './resources/MessagingShortCodes';
+import {MessagingTollfreeVerification} from './resources/MessagingTollfreeVerification';
+import {MessagingUrlDomains} from './resources/MessagingUrlDomains';
+import {MobileNetworkOperators} from './resources/MobileNetworkOperators';
+import {MobileOperatorNetworks} from './resources/MobileOperatorNetworks';
+import {Networks} from './resources/Networks';
+import {NumberBackgroundJobs} from './resources/NumberBackgroundJobs';
+import {NumberLookup} from './resources/NumberLookup';
+import {NumberOrderDocuments} from './resources/NumberOrderDocuments';
+import {NumberOrders} from './resources/NumberOrders';
+import {NumberPortouts} from './resources/NumberPortouts';
+import {NumberReservations} from './resources/NumberReservations';
+import {NumbersFeatures} from './resources/NumbersFeatures';
+import {Object as ObjectResource} from './resources/Object';
+import {OtaUpdates} from './resources/OtaUpdates';
+import {Outbound} from './resources/Outbound';
+import {OutboundVoiceProfiles} from './resources/OutboundVoiceProfiles';
+import {Payment} from './resources/Payment';
+import {PhoneNumberAssignmentByProfile} from './resources/PhoneNumberAssignmentByProfile';
+import {PhoneNumberBlockOrders} from './resources/PhoneNumberBlockOrders';
+import {PhoneNumberBlocksBackgroundJobs} from './resources/PhoneNumberBlocksBackgroundJobs';
+import {PhoneNumberCampaigns} from './resources/PhoneNumberCampaigns';
+import {PhoneNumberOrderDocuments} from './resources/PhoneNumberOrderDocuments';
+import {PhoneNumberRegulatoryRequirements} from './resources/PhoneNumberRegulatoryRequirements';
+import {PhoneNumbers} from './resources/PhoneNumbers';
+import {PhoneNumberSearch} from './resources/PhoneNumberSearch';
+import {PhoneNumbersInboundChannels} from './resources/PhoneNumbersInboundChannels';
+import {PhoneNumbersMessaging} from './resources/PhoneNumbersMessaging';
+import {PhoneNumbersVoice} from './resources/PhoneNumbersVoice';
+import {PortabilityChecks} from './resources/PortabilityChecks';
+import {PortingOrders} from './resources/PortingOrders';
+import {PortingPhoneNumbers} from './resources/PortingPhoneNumbers';
+import {PortoutRequests} from './resources/PortoutRequests';
+import {PresignedObjectUrls} from './resources/PresignedObjectUrls';
+import {PrivateWirelessGateways} from './resources/PrivateWirelessGateways';
+import {PublicInternetGateways} from './resources/PublicInternetGateways';
+import {PublicKey} from './resources/PublicKey';
+import {PushCredentials} from './resources/PushCredentials';
+import {Queues} from './resources/Queues';
+import {RecordingsCommands} from './resources/RecordingsCommands';
+import {Regions} from './resources/Regions';
+import {RegisterCall} from './resources/RegisterCall';
+import {RegulatoryRequirements} from './resources/RegulatoryRequirements';
+import {Reporting} from './resources/Reporting';
+import {Reports} from './resources/Reports';
+import {ReportsMdrs} from './resources/ReportsMdrs';
+import {Requirements} from './resources/Requirements';
+import {RequirementTypes} from './resources/RequirementTypes';
+import {RoomCompositions} from './resources/RoomCompositions';
+import {RoomParticipants} from './resources/RoomParticipants';
+import {RoomRecordings} from './resources/RoomRecordings';
+import {Rooms} from './resources/Rooms';
+import {RoomsClientToken} from './resources/RoomsClientToken';
+import {RoomsClientTokens} from './resources/RoomsClientTokens';
+import {RoomSessions} from './resources/RoomSessions';
+import {SharedCampaigns} from './resources/SharedCampaigns';
+import {ShortCodes} from './resources/ShortCodes';
+import {SimCardActions} from './resources/SimCardActions';
+import {SimCardGroupActions} from './resources/SimCardGroupActions';
+import {SimCardGroups} from './resources/SimCardGroups';
+import {SimCardOrders} from './resources/SimCardOrders';
+import {SimCards} from './resources/SimCards';
+import {SslCertificates} from './resources/SslCertificates';
+import {TelephonyCredentials} from './resources/TelephonyCredentials';
+import {TeXMLApplication} from './resources/TeXMLApplication';
+import {UpdateClientState} from './resources/UpdateClientState';
+import {Verifications} from './resources/Verifications';
+import {VerifiedCallsDisplayProfile} from './resources/VerifiedCallsDisplayProfile';
+import {VerifiedCallsDisplayProfiles} from './resources/VerifiedCallsDisplayProfiles';
+import {VerifiedNumbers} from './resources/VerifiedNumbers';
+import {Verify} from './resources/Verify';
+import {VerifyProfiles} from './resources/VerifyProfiles';
+import {VirtualCrossConnects} from './resources/VirtualCrossConnects';
+import {WdrDetailReports} from './resources/WdrDetailReports';
+import {Webhooks as WebhooksResource} from './resources/Webhooks';
+import {WireguardInterfaces} from './resources/WireguardInterfaces';
+import {WirelessDetailRecordReports} from './resources/WirelessDetailRecordReports';
 //
 
 import TelnyxResource from './TelnyxResource';
@@ -43,11 +192,160 @@ export function createTelnyx() {
 
   const APP_INFO_PROPERTIES = ['name', 'version', 'url', 'partner_id'];
 
-  // TODO: convert other resources to ts
+  // #region Resources Definition
   const resources = {
+    AccessIpAddress,
+    AccessIpRanges,
+    AccessTokens,
+    Actions,
+    ActionsSimCards,
+    ActivateDeactivateBulkCredentials,
+    Addresses,
+    AdvancedOptinoptout,
+    AI,
+    AuthenticationProviders,
+    AvailablePhoneNumbers,
     Balance,
+    Billing,
+    BillingGroups,
+    Brands,
+    Bucket,
+    BucketUsage,
+    BulkCreation,
+    BulkCredentials,
+    BulkPhoneNumberCampaigns,
+    BulkPhoneNumberOperations,
+    BulkSoleProprietorCreation,
+    BulkTelephonyCredentials,
+    BusinessIdentity,
+    CallControlApplications,
+    CallEvents,
+    CallInformation,
+    CallRecordings,
+    Calls,
+    Campaign,
+    CampaignBuilder,
+    CdrUsageReports,
+    Channelzones,
+    ClientStateUpdate,
+    Conferences,
+    Connections,
+    CredentialConnections,
+    Credentials,
+    CsvDownloads,
+    CustomerServiceRecord,
+    Debugging,
+    DetailRecords,
+    DialogflowIntegration,
+    DocumentLinks,
+    Documents,
+    DynamicEmergency,
+    DynamicEmergencyAddresses,
+    DynamicEmergencyEndpoints,
+    Events,
+    ExternalConnections,
+    FaxApplications,
+    Faxes,
+    FqdnConnections,
+    Fqdns,
+    GlobalIps,
+    InboundChannels,
+    InventoryCoverage,
+    InventoryLevel,
+    IpAddresses,
+    IpConnections,
+    IpRanges,
+    Ips,
+    ManagedAccounts,
+    MdrDetailReports,
+    MdrUsageReports,
+    MediaStorageApi,
+    Messages,
+    MessagesAlphanumericSenderId,
+    MessagingHostedNumber,
+    MessagingHostedNumberOrders,
+    MessagingHostedNumbers,
+    MessagingPhoneNumbers,
     MessagingProfileMetrics,
     MessagingProfiles,
+    MessagingSenderIds,
+    MessagingShortCodes,
+    MessagingTollfreeVerification,
+    MessagingUrlDomains,
+    MobileNetworkOperators,
+    MobileOperatorNetworks,
+    Networks,
+    NumberBackgroundJobs,
+    NumberLookup,
+    NumberOrderDocuments,
+    NumberOrders,
+    NumberPortouts,
+    NumberReservations,
+    NumbersFeatures,
+    Object: ObjectResource,
+    OtaUpdates,
+    Outbound,
+    OutboundVoiceProfiles,
+    Payment,
+    PhoneNumberAssignmentByProfile,
+    PhoneNumberBlockOrders,
+    PhoneNumberBlocksBackgroundJobs,
+    PhoneNumberCampaigns,
+    PhoneNumberOrderDocuments,
+    PhoneNumberRegulatoryRequirements,
+    PhoneNumbers,
+    PhoneNumberSearch,
+    PhoneNumbersInboundChannels,
+    PhoneNumbersMessaging,
+    PhoneNumbersVoice,
+    PortabilityChecks,
+    PortingOrders,
+    PortingPhoneNumbers,
+    PortoutRequests,
+    PresignedObjectUrls,
+    PrivateWirelessGateways,
+    PublicInternetGateways,
+    PublicKey,
+    PushCredentials,
+    Queues,
+    RecordingsCommands,
+    Regions,
+    RegisterCall,
+    RegulatoryRequirements,
+    Reporting,
+    Reports,
+    ReportsMdrs,
+    Requirements,
+    RequirementTypes,
+    RoomCompositions,
+    RoomParticipants,
+    RoomRecordings,
+    Rooms,
+    RoomsClientToken,
+    RoomsClientTokens,
+    RoomSessions,
+    SharedCampaigns,
+    ShortCodes,
+    SimCardActions,
+    SimCardGroupActions,
+    SimCardGroups,
+    SimCardOrders,
+    SimCards,
+    SslCertificates,
+    TelephonyCredentials,
+    TeXMLApplication,
+    UpdateClientState,
+    Verifications,
+    VerifiedCallsDisplayProfile,
+    VerifiedCallsDisplayProfiles,
+    VerifiedNumbers,
+    Verify,
+    VerifyProfiles,
+    VirtualCrossConnects,
+    WdrDetailReports,
+    Webhooks: WebhooksResource,
+    WireguardInterfaces,
+    WirelessDetailRecordReports,
   };
   //